requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
sentence-transformers==3.3.1
faiss-cpu==1.9.0
openai==1.58.1
numpy==1.26.4
python-dotenv